[TE] Support ScaleFabric SHCA NICs
Add USE_SHCA build option for ScaleFabric SHCA NICs, which use
extended 17-bit LIDs instead of standard 16-bit IB LIDs. When enabled,
convert LIDs via shca_17b_types.h helpers and widen lid fields to
uint32_t across Transfer Engine and TENT RDMA paths.

Gate yalantinglibs IBV RPC (YLT_ENABLE_IBV) behind USE_SHCA=OFF so SHCA
builds skip ib_socket-based RDMA RPC and fall back to TCP with a warning.

Adjust dependencies.sh to detect SHCA headers and skip conflicting system
RDMA packages (libibverbs-dev on Debian, rdma-core-devel on RHEL).

+++ mooncake-transfer-engine/include/transport/rdma_twosided/ctrl_channel.h
@@ -95,7 +95,7 @@ class CtrlChannel {
     void destroyResources();
     int postRecv(size_t idx);
     int repostAllRecvs();
-    int connectQp(const std::string &peer_gid, uint16_t peer_lid,
+    int connectQp(const std::string &peer_gid, uint32_t peer_lid,
                   uint32_t peer_qp_num);
     void dispatchRecvPayload(const uint8_t *data, size_t byte_len);
     void handleSendComplete();
